+++ event-sourcing/subscription-manager/src/test/java/uk/gov/justice/services/event/sourcing/subscription/manager/cdi/factories/NewSubscriptionManagerFactoryTest.java
@@ -5,6 +5,7 @@ import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static uk.gov.justice.services.test.utils.core.reflection.ReflectionUtil.getValueOfField;
 
+import uk.gov.justice.services.common.configuration.subscription.pull.EventPullConfiguration;
 import uk.gov.justice.services.event.sourcing.subscription.manager.NewSubscriptionManager;
 import uk.gov.justice.services.event.sourcing.subscription.manager.NewSubscriptionManagerDelegate;
 import uk.gov.justice.services.subscription.SubscriptionManager;
@@ -21,6 +22,9 @@ public class NewSubscriptionManagerFactoryTest {
     @Mock
     private NewSubscriptionManagerDelegate newSubscriptionManagerDelegate;
 
+    @Mock
+    private EventPullConfiguration eventPullConfiguration;
+
     @InjectMocks
     private NewSubscriptionManagerFactory newSubscriptionManagerFactory;
 
@@ -34,6 +38,7 @@ public class NewSubscriptionManagerFactoryTest {
         assertThat(subscriptionManager, is(instanceOf(NewSubscriptionManager.class)));
 
         assertThat(getValueOfField(subscriptionManager, "newSubscriptionManagerDelegate", NewSubscriptionManagerDelegate.class), is(newSubscriptionManagerDelegate));
+        assertThat(getValueOfField(subscriptionManager, "eventPullConfiguration", EventPullConfiguration.class), is(eventPullConfiguration));
         assertThat(getValueOfField(subscriptionManager, "componentName", String.class), is(componentName));
     }
 }